Outbound Routes: Uploaded csv-file is not deemed valid csv (Fixed)
In special circumstances a uploaded csv-file with dial patterns is not recognized as valid, because some browsers send another mime-type.
This PR removes the check if a csv-file was uploaded based on the mime type and adds a check if the uploaded file is in the correct format for FreePBX 16.

Fixed missing declaration of variables

Bugfix FreePBX/issue-tracker#678